Moved coordinates functions to a separated file
