+++ relion/convert/__init__.py
@@ -26,6 +26,7 @@
 
 from .convert_utils import *
 from .convert_deprecated import *
+from .convert_coordinates import *
 from .dataimport import *
 import relion
 import math
